requirements.txt
# Plastic Waste Detection for Recycling Automation
# Python Dependencies

# Core ML & Computer Vision
ultralytics==8.0.200  # YOLOv8
torch==2.1.1  # PyTorch (CPU, change to torch with CUDA for GPU)
torchvision==0.16.1
opencv-python==4.8.1.78  # OpenCV

# Data Processing & Augmentation
numpy==1.24.3
Pillow==10.0.0
pandas==2.0.3
scikit-learn==1.3.0
albumentations==1.3.0  # Image augmentation

# Web Scraping
icrawler==0.6.10  # Image crawler for web scraping (compatible release)
requests==2.31.0
beautifulsoup4==4.12.2

# API & Web Framework
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
starlette==0.27.0

# Configuration Management
PyYAML==6.0.1

# Data Visualization & Metrics
matplotlib==3.8.1
seaborn==0.13.0

# Logging & Monitoring
python-json-logger==2.0.7

# Development & Testing
pytest==7.4.3
black==23.11.0  # Code formatter
flake8==6.1.0  # Linter
mypy==1.7.1  # Type checker

# Optional: GPU Support (uncomment if using CUDA)
# torch-cuda==12.1  # For NVIDIA GPUs
# torch-cuda-tools==12.1
